chore(tests): move results_file_dumper
Move it under performance since that's where it is used.
